// src/api/generated/model/updateScamAlertDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Honest Bangkok API
 * The Honest Bangkok API for locals and tourists to share tips, prices, and scam alerts.
 * OpenAPI spec version: 1.0
 */

export interface UpdateScamAlertDto {
  /** City ID */
  cityId?: string;
  /** Category ID */
  categoryId?: string;
  /** Scam name */
  scamName?: string;
  /** Scam description */
  description?: string;
  /** Prevention tip */
  preventionTip?: string;
  /** Scam alert image (JPG, PNG, WebP - max 10MB, optional) */
  image?: Blob;
}